+++ .eslintcache
@@ -1,0 +1,1 @@
+[{"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/index.tsx":"1","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/reportWebVitals.ts":"2","/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/App/App.tsx":"3"},{"size":504,"mtime":1609901260491,"results":"4","hashOfConfig":"5"},{"size":425,"mtime":1609899106504,"results":"6","hashOfConfig":"5"},{"size":647,"mtime":1609901256880,"results":"7","hashOfConfig":"5"},{"filePath":"8","messages":"9","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"17bwe72",{"filePath":"10","messages":"11","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"12","messages":"13","errorCount":2,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/index.tsx",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/reportWebVitals.ts",[],"/Users/lopez.4.joseph/turing/mod_3/covid_comparisons/src/App/App.tsx",["14","15"],{"ruleId":"16","severity":2,"message":"17","line":26,"column":14,"nodeType":"18","endLine":26,"endColumn":22},{"ruleId":"16","severity":2,"message":"19","line":30,"column":14,"nodeType":"18","endLine":30,"endColumn":22},"react/jsx-no-undef","'Dropdown' is not defined.","JSXIdentifier","'USAState' is not defined."]
