--- src/App.tsx
@@ -1,21 +1,0 @@
-import React, {Component} from 'react';
-import './App.css';
-
-interface Props { 
-
-}
-
-interface State {
-  selectedUSAState: {};
-}
-
-class App extends Component<Props,State> {
-  constructor(props: Props) {
-    super(props);
-    this.state = {
-      selectedUSAState: {}
-    }
-  }
-}
-
-export default App;
