+++ src/ComparisonDetails/ComparisonDetails.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const ComparisonDetails = () => {
+
+}
+
+export default ComparisonDetails;
