+++ src/Nav/Nav.tsx
@@ -1,0 +1,7 @@
+import React from 'react';
+
+const Nav = () => {
+
+}
+
+export default Nav;
