+++ src/USAState/USAState.tsx
@@ -1,0 +1,6 @@
+import React from 'react';
+
+const USAState = () => {
+}
+
+export default USAState;
